+++ libs/xml/src/XmlNode.ts
@@ -1,15 +1,1 @@
-
-/**
- * XML node
- *
- * @public
- */
-export type XmlNode = {
-	nodeName: string;
-	nodeValue: string | null;
-	attributes: Record<string, string>;
-	childNodes: XmlNode[];
-	prefix?: string | null;
-	localName?: string;
-	parentElement?: XmlNode | null;
-};
+export type { XmlNode } from '@svta/cml-utils'