+++ libs/xml/CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [1.1.2] - 2026-02-04
+
+### Changed
+
+- Re-export `XmlNode` type from `@svta/cml-utils`
+- Update `@svta/cml-utils` to 1.5.0
+
 ## [1.1.1] - 2026-02-03
 
 ### Changed
@@ -44,7 +51,8 @@ and this project adheres to
 - Convert to mono-repo ([#238](https://github.com/streaming-video-technology-alliance/common-media-library/issues/238))
 - Produce single bundled export for each package ([#260](https://github.com/streaming-video-technology-alliance/common-media-library/issues/260))
 
-[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/xml-v1.1.1...HEAD
+[Unreleased]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/xml-v1.1.2...HEAD
+[1.1.2]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/xml-v1.1.1...xml-v1.1.2
 [1.1.1]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/xml-v1.1.0...xml-v1.1.1
 [1.1.0]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/xml-v1.0.2...xml-v1.1.0
 [1.0.2]: https://github.com/streaming-video-technology-alliance/common-media-library/compare/xml-v1.0.1...xml-v1.0.2